Campos montos

+++ src/main/java/com/banvenez/ast/dto/farmacia/RegistrarInventarioDto.java
@@ -2,6 +2,8 @@ package com.banvenez.ast.dto.farmacia;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 
 
@@ -10,9 +12,9 @@ public class RegistrarInventarioDto {
     private String  codarticulo;
     private String  descripcion;
     private Integer  codtipo;
-    private Integer cantidad;
-    private Double  preciocosto;
-    private Double  precioventa;
+    private Long  cantidad;
+    private BigDecimal   preciocosto;
+    private BigDecimal   precioventa;
     private Double  descuento;
     private Integer volumen;
     private String  status;
